+++ Changelog.md
@@ -11,6 +11,7 @@
 - Changed minimal Enum requirement to v4.1
 - Upgraded to Konekt Address, Customer and User modules to v3
 - Removed the Vanilo v2 `Framework` namespace compatibility layer
+- Added the `currency` field to the orders table
 
 ## 3.x Series