+++ src/Order/Changelog.md
@@ -10,6 +10,7 @@
 - Dropped Enum v3 Support
 - Changed minimal Enum requirement to v4.1
 - Upgraded to Konekt Address and User modules to v3
+- Added the `currency` field to the orders table
 
 ## 3.x Series